Use powerlaw for insert/delete distribution in generate_data tests
This change makes sure that inserts/deletes only happen on the "most
recent" data. It also changes the logic to only load and write the slice
of the Dataset which is actually impacted which should improve the
performance.

I also refactored the duplicated code between the versioned and regular
test case out into a common base class.

Note that I cannot verify that the versioned tests still succeed since
they no longer work against master, unfortunately.
